Changes related to new buttons Previous & Next chapter, TinyMCE, several messages & right part of the navbar on public pages  (in case of open session)
New feature of navigation with buttons Previous chapter & Next chapter. Addition of TinyMCE with some options & plugins for the chapters forms. Addition of messages : request for confirmation of the deletion of  a chapter, info for reporting a comment & info displayed on the page Visualisation de chapitre in case of a too large photo. Right part added on the public navbar in case of open session. Little simplification of code.

+++ view/backend/postsMgtAdmView.php
@@ -44,7 +44,7 @@
                         <td class="<?= ($data['publicationDate']> date("Y-m-d H:i:s") OR empty($data['publicationDate_fr']))?'orange':' ' ?>" > <em> <?=(!empty($data['publicationDate_fr']) ) ? $data['publicationDate_fr'] : 'Date à déterminer' ?></em> </td>
                         <td ><em><a  href="index.php?action=viewPostAdm&amp;id=<?= $data['id'] ?>"><span class="glyphicon glyphicon-eye-open big"></span></</a></em></td>
                         <td ><em><a href="index.php?action=rectifyFormPost&amp;id=<?= $data['id'] ?>"><span class="glyphicon glyphicon-pencil big"></span></a></em></td>
-                        <td ><em><a href="index.php?action=deletePost&amp;id=<?= $data['id'] ?>">
+                        <td ><em><a href="index.php?action=deletePost&amp;id=<?= $data['id'] ?>" onclick="return confirm('Êtes-vous certain de vouloir effacer définitivement ce chapitre ?')">
                         <span class="glyphicon glyphicon-remove-sign big"></span></a></em></td>
                     </tr>
                     <?php
